moved UserCard styles out of component
